% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{setup}
\alias{setup}
\title{Setup folder structure for analysis}
\usage{
setup(path = ".")
}
\arguments{
\item{path}{character string, specifies where to make folders. Defaults to current directory}
}
\value{

}
\description{
setup(path) checks if the following folders exist at path. If not, setup
makes them:
}
\details{
* data - for any raw data files
* tmp - for any temporary files (e.g. processed data, models, ...)
* out - for any outputs (e.g. images, tables for sharing, ...)
}
\examples{
setup()
}
